requirements.txt
absl-py==0.9.0
astroid==2.4.2
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
cycler==0.10.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
humanfriendly==8.2
idna==2.9
isort==4.3.21
joblib==0.15.1
Keras==2.4.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.7.0
toml==0.10.1
torch==1.5.1
tqdm==4.46.1
transformers==2.11.0
urllib3==1.25.9
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
